Search button makes ajax call and appends json response to div
This isn't complete, but it took me a while to do so I'm
committing. At this point, when you press the search button it
successfully makes an ajax request to event#search, which
responds with json and that's appended to the div.

I now need to respond with the events and replace the events
with the results.